[CLEAN]

+++ src/main.c
@@ -554,7 +554,7 @@ void midi_write_multiple_euclidean(t_music_data *music_data, t_sensors *sensors_
 		}
 	}
 
-	if (delta_shift != (uint32_t)map_number((uint32_t)sensors_data->spectro_current, 0, 33535, 0, 10))
+	if (delta_shift != (uint16_t)map_number((uint32_t)sensors_data->spectro_current, 0, 33535, 0, 10))
 	{
 
 		int16_t tmp = (uint32_t)map_number((uint32_t)sensors_data->spectro_current, 0, 33535, 0, 10) - delta_shift;
